requirements.txt
flask
torch
torchvision
pillow
numpy
pandas
gdown
gunicorn
